Moved engine functions into engine.py
gameshell.py now only contains IO related functionality
